# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatePresetRequest": {
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/ProcessingOptions"
          }
        },
        "required": [
          "name",
          "options"
        ],
        "title": "CreatePresetRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelBenchmarkReport": {
        "properties": {
          "benchmark_dir": {
            "title": "Benchmark Dir",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "log_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Path"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ModelBenchmarkResult"
            },
            "title": "Results",
            "type": "array"
          },
          "sample_dir": {
            "title": "Sample Dir",
            "type": "string"
          },
          "task": {
            "enum": [
              "cutout",
              "enhance",
              "tagging",
              "ocr"
            ],
            "title": "Task",
            "type": "string"
          },
          "total_images": {
            "title": "Total Images",
            "type": "integer"
          },
          "total_models": {
            "title": "Total Models",
            "type": "integer"
          }
        },
        "required": [
          "task",
          "sample_dir",
          "benchmark_dir",
          "generated_at",
          "total_models",
          "total_images",
          "results"
        ],
        "title": "ModelBenchmarkReport",
        "type": "object"
      },
...
      "ModelBenchmarkStatus": {
        "properties": {
          "benchmark_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Dir"
          },
          "completed_images": {
            "default": 0,
            "title": "Completed Images",
            "type": "integer"
          },
          "completed_models": {
            "default": 0,
            "title": "Completed Models",
            "type": "integer"
          },
          "current_image_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Image Name"
          },
          "current_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Model Id"
          },
          "current_model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Model Name"
          },
          "current_model_started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Model Started At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "log_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Path"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelBenchmarkReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "sample_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Dir"
          },
          "state": {
            "enum": [
              "idle",
              "running",
              "done",
              "failed"
            ],
            "title": "State",
            "type": "string"
          },
          "task": {
            "default": "cutout",
            "enum": [
              "cutout",
              "enhance",
              "tagging",
              "ocr"
            ],
            "title": "Task",
            "type": "string"
          },
          "total_images": {
            "default": 0,
            "title": "Total Images",
            "type": "integer"
          },
          "total_models": {
            "default": 0,
            "title": "Total Models",
            "type": "integer"
          }
        },
        "required": [
          "state"
        ],
        "title": "ModelBenchmarkStatus",
        "type": "object"
      },
      "ModelCatalogItem": {
        "properties": {
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "engine": {
            "enum": [
              "rembg",
              "realesrgan",
              "wdtagger",
              "tesseract"
            ],
            "title": "Engine",
            "type": "string"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "included_by_default": {
            "default": false,
            "title": "Included By Default",
            "type": "boolean"
          },
          "install_state": {
            "enum": [
              "not_installed",
              "downloading",
              "installed",
              "failed"
            ],
            "title": "Install State",
            "type": "string"
          },
          "installable": {
            "default": true,
            "title": "Installable",
            "type": "boolean"
          },
          "installed": {
            "title": "Installed",
            "type": "boolean"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "local_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Path"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "runtime_ready": {
            "default": false,
            "title": "Runtime Ready",
            "type": "boolean"
          },
          "size_mb": {
            "title": "Size Mb",
            "type": "number"
          },
          "status_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Note"
          },
          "task": {
            "enum": [
              "cutout",
              "enhance",
              "tagging",
              "ocr"
            ],
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task",
          "engine",
          "name",
          "description",
          "size_mb",
          "installed",
          "install_state"
        ],
        "title": "ModelCatalogItem",
        "type": "object"
      },
      "ModelStatusResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "install_state": {
            "enum": [
              "not_installed",
              "downloading",
              "installed",
              "failed"
            ],
            "title": "Install State",
            "type": "string"
          },
          "installed": {
            "title": "Installed",
            "type": "boolean"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "local_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Path"
          }
        },
        "required": [
          "id",
          "installed",
          "install_state"
        ],
        "title": "ModelStatusResponse",
        "type": "object"
      },
      "ModelStorageConfig": {
        "properties": {
          "default_root_dir": {
            "title": "Default Root Dir",
            "type": "string"
          },
          "rembg_dir": {
            "title": "Rembg Dir",
            "type": "string"
          },
          "root_dir": {
            "title": "Root Dir",
            "type": "string"
          },
          "using_custom_root": {
            "title": "Using Custom Root",
            "type": "boolean"
          }
        },
        "required": [
          "root_dir",
          "rembg_dir",
          "default_root_dir",
          "using_custom_root"
        ],
        "title": "ModelStorageConfig",
        "type": "object"
      },
      "PresetItem": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_builtin": {
            "default": false,
            "title": "Is Builtin",
            "type": "boolean"
          },
          "is_editable": {
            "default": true,
            "title": "Is Editable",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/ProcessingOptions"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "options",
          "created_at",
          "updated_at"
        ],
        "title": "PresetItem",
        "type": "object"
      },
...
      "RunModelBenchmarkRequest": {
        "properties": {
          "model_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Ids"
          },
          "sample_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Dir"
          },
          "task": {
            "default": "cutout",
            "enum": [
              "cutout",
              "enhance",
              "tagging",
              "ocr"
            ],
            "title": "Task",
            "type": "string"
          }
        },
        "title": "RunModelBenchmarkRequest",
        "type": "object"
      },
      "UpdateModelStorageRequest": {
        "properties": {
          "migration_mode": {
            "default": "move",
            "enum": [
              "move",
              "delete",
              "ignore"
            ],
            "title": "Migration Mode",
            "type": "string"
          },
          "root_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Dir"
          }
        },
        "title": "UpdateModelStorageRequest",
        "type": "object"
      },
      "UpdatePresetRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProcessingOptions"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "UpdatePresetRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ClearCut",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/models/benchmark/run": {
      "post": {
        "operationId": "run_model_benchmark_models_benchmark_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunModelBenchmarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelBenchmarkStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Model Benchmark",
        "tags": [
          "models"
        ]
      }
    },
...
    "/models/catalog": {
      "get": {
        "operationId": "list_model_catalog_models_catalog_get",
        "parameters": [
          {
            "in": "query",
            "name": "task",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelCatalogItem"
                  },
                  "title": "Response List Model Catalog Models Catalog Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Model Catalog",
        "tags": [
          "models"
        ]
      }
    },
    "/models/config": {
      "get": {
        "operationId": "get_model_storage_config_models_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelStorageConfig"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Model Storage Config",
        "tags": [
          "models"
        ]
      },
      "patch": {
        "operationId": "update_model_storage_config_models_config_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModelStorageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelStorageConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Model Storage Config",
        "tags": [
          "models"
        ]
      }
    },
    "/models/refresh": {
      "post": {
        "operationId": "refresh_model_catalog_models_refresh_post",
        "parameters": [
          {
            "in": "query",
            "name": "task",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelCatalogItem"
                  },
                  "title": "Response Refresh Model Catalog Models Refresh Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Model Catalog",
        "tags": [
          "models"
        ]
      }
    },
    "/models/{model_id}": {
      "delete": {
        "operationId": "delete_model_models__model_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Model Models  Model Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Model",
        "tags": [
          "models"
        ]
      }
    },
    "/models/{model_id}/install": {
      "post": {
        "operationId": "install_model_models__model_id__install_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Install Model",
        "tags": [
          "models"
        ]
      }
    },
    "/models/{model_id}/status": {
      "get": {
        "operationId": "get_model_status_models__model_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Model Status",
        "tags": [
          "models"
        ]
      }
    },
    "/presets": {
      "get": {
        "operationId": "list_presets_presets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PresetItem"
                  },
                  "title": "Response List Presets Presets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Presets",
        "tags": [
          "presets"
        ]
      },
      "post": {
        "operationId": "create_preset_presets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePresetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Preset",
        "tags": [
          "presets"
        ]
      }
    },
    "/presets/{preset_id}": {
      "delete": {
        "operationId": "delete_preset_presets__preset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "preset_id",
            "required": true,
            "schema": {
              "title": "Preset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Preset Presets  Preset Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Preset",
        "tags": [
          "presets"
        ]
      },
      "patch": {
        "operationId": "update_preset_presets__preset_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "preset_id",
            "required": true,
            "schema": {
              "title": "Preset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePresetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresetItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Preset",
        "tags": [
          "presets"
        ]
      }
    },
...
  }
}